feature: make exp-reduce works on families
Problem: The rollup of information from single ensembles
to a "family" or "experiment" uses reduce_drm_sets, and
the Makefile (make ... exp-reduce) passes in the auto-generated
name of an index file. For families, this file is not there,
but it's inessential.

Solution: Make the error on missing index file into a warning.

Testing: Ran some demo families with make exp-reduce.

Origin: