feature: compile ensemble numbers for families
Problem: the family index.html page had no indication of
how many ensembles were part of the family, and when it
was last run.

Solution: read the reduce-info.csv file and add a table
row (footer) to put a total line in. Improved the tables
to have a <tfoot> that is not sorted with the <tbody>.

Testing: running make ... html-only

Origin: mike