feature: add simtime to reduction and html display
Problem: We need a way to identify older simulations. The runtime isn't
ideal for that because sometimes we re-run reductions for no
particular reason, or by mistake.

Solution: Add the last simulation time (timestamp of DRM folder) to
the data reduction and to the HTML display. The rollup for
Experiments and Families takes the latest date within the
family of ensembles. Additionally, the simtime had to be
retrospectively added to the reduce-info.csv files across
all ensembles, without altering the timestamp of the file
(or else the graphics, etc., would seem out-of-date!).

Testing: Trial runs.

Origin: turmon