feature: ensemble families are rolled up correctly
Problem: We support grouping of sets of ensembles into
a "Family" which is basically a subdirectory. We want
to support Families containing Families, or Experiments,
etc. This required some changes to the html-summary program,
and to the reduce_drm_sets program.

Solution: The indexing was made more systematic. Families
and Experiments of arbitrary depth are supported.

Testing: extensive testing with example families and
experiments and ensembles.

Origin: turmon