feature: new utility tabulates drm observations
Problem: Existing util/drm-grep.py could only write to CSV, but
some DRM entries, especially vectors of variable length, do not
translate to CSV.

Solution: Generalization.
1/ Add ability to evaluate an arbitrary python expression
in the context of each DRM observation, and output the result
of that evaluation.
2/ Add ability to output to JSON as well as to CSV. JSON can
be read into Matlab or Python.
Additionally, because the expressions can get more complex,
the ability to load the attribute/expression list from a file
was added. This is a cheap way to get more flexible reduction
capability for some special one-offs.

Testing:

Origin: